requirements.txt
streamlit==1.28.1
requests==2.31.0
pypdf==3.17.1
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
sqlite3
python-dotenv==1.0.0
faiss-cpu==1.7.4
scikit-learn==1.3.0 